+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Update dropdown styling to accomodate long style names
+- Add layer id to feature popup
 - Fix bug for branch dropdown option while polling a style
 
 ## 0.13.0